Tg 28 (#20)
* organization

* favoriting is now wired up

favoriting is now wired up.  it refreshes the entries to keep everything synced up.  this is perfectly fine when youre inside the FM but when you are on entryDetails, there is a flashing which is kind of annoying.

* foundational bird tagging has been added

The user can now toggle foundation/non foundation on and off within the FM & entryDetails pages.  The goal of this is to aid the user in sorting and filtering.