some refactor and prep for edit mode
Because edit mode will tether the actual values of the entry to the components in a more two-way manner, props will no longer work.  Rather, they would work but it would only be one way, which as i type this out may be exactly what we want to keep things top down and only update once a query is fired... more thinking.