Tg 55 (#33)
* some refactor and prep for edit mode

Because edit mode will tether the actual values of the entry to the components in a more two-way manner, props will no longer work.  Rather, they would work but it would only be one way, which as i type this out may be exactly what we want to keep things top down and only update once a query is fired... more thinking.

* added edit toggle

working through reactivity and the right way to deal with these changes.  either slots or v-ifs. i hate both and they both arent doing the job very well either.

* Edit mode is mostly wired up to chips

had to do some various refactoring to make this more smooth.  Still a non-props extraneous value floating around but it's been there for a while.  Going to move onto adding save button.

* rewired chars a bit

* editing and saving works

need to clean up

* solved vue warning

props passed in from router need to be defined on the page in which renders them first.

* removed watcher and other wasteful logic

Was able to remove onmounted, watcher, and a number of other logic which was overcomplicated this.  It now updates the chars only and handles it gracefully using v-model.

* chips are now styled mostly correctly

cutting here to work on star rating and other attributes.

* cutting here for megan

* major cleanup.