Tg 48 (#38)
* added more seamless snacks

snackbar now stacks more nicely.  styled snackbar. added icons. reworked some of the onSnapshot logic to not update entries which are not touched. updated messages to use actual onSnapshot(change) details for more precision.

* restyled FM page to allow more space

* adding styles to icons

* tg-48 confirmation dialog semi wired up

converted it to a small promise to not have to interrupt or break the actions.  cancel works. confirmation doesnt yet.  it also needs styled.

* confirmation dialog is wired up

need to do a little styling

* tg-48 prepping for pr