Tg 68 (#41)
* tg-68 added notes section

added notes to card using new v-model pattern.  also replaced it on the add entry form. 2 remaining components to pull out of schema before it can be completely removed.

* tg-68 resolved some bugs caused by removing defaults during last PR

need to break out notes into 2 groups. archived and active. both need to be objects since they will save the date and time also.

* tg-68 notes handled on saving side

need to rewire up the created/add entry portion of the notes area.  Need to also fix the breeds picker that i previously broke,.

* tg-68 resolved issue where clearing form would break notes

* tg-68 various styling updates and QoL improvements

forms and cards updated with appropriate spacers and headings.  Spacing on cards minor adjustment.

* tg-68 some minor styling for mobile

* tg-68 wiring up data and handling for archived notes displays

adding userName, photoUrl, and content to the card.  this will allow for a more appealing display but requires a but of massaging and updating

* tg-68 finished styling for archived comments

also restyled and realigned some spacing on the main card.