Tg 54 (#46)
* tg-54 added capacitor, wired up mobile app baseline

* tg-54 saving prog

* tg-54 getting closer

I believe the scrollBehaviors are fighting with one another which is causing a quick pause when you transition from scrolling up to down or vice versa.

* tg-54 cutting here fixed jittering

Still need to figure out how to keep the mother fucking stupid main scroll area to not go underneath the top bar.

* tg-54 leaving it like this. works good enough

i have a feeling this is going to break other scrolling to top and stopping things, like filters staying at the top.  The wrong things are scrolling 'i think'.  Can confirm later.

* tg-54 tg-54 added pull to refresh for mobile

* tg-54 tg-54 styling and fighting with the drag refresh.

need to finish up some styling and wiring for the drag refresh but it's close.  It's important to note the possibility of simply refreshing the data vs app during a pull/drag.  this is very different from reloading the entire app and offers some unique pros.

* tg-54 wired up pull to refresh with clearing of store and rehydration of entries within the app.

need to work on minor todos

* tg-54 resolved bug where bottom padding was being wrongly applied

* tg-54 resolved bug where skeleton loaders widths were not being calculated accurately

* tg-54 cleanup

* tg-54 cleanup